Modificación de tipo de dato en la BD
El campo calificacion de AluEntregaAsi estaba como decimal(4,0). Se
cambia a float.

+++ ET3.sql
@@ -34,7 +34,7 @@ CREATE TABLE IF NOT EXISTS `AluEntregaTra` (
   `fechaEntrega` date NOT NULL,
   `observaciones` varchar(256) DEFAULT NULL,
   `titulo` varchar(80) NOT NULL,
-  `calificacion` decimal(4,0) DEFAULT NULL,
+  `calificacion` float DEFAULT NULL,
   `portfolio` enum('T','F') NOT NULL DEFAULT 'F'
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
